+++ src/sbbs3/getstr.cpp
@@ -50,7 +50,7 @@ size_t sbbs_t::getstr(char *strout, size_t maxlen, long mode, const str_list_t h
                     /* x&z=misc */
 	char	str1[256],str2[256],undo[256];
     uchar	ch;
-	uchar	atr;
+	uint	atr;
 	int		hidx = -1;
 
 	long term = term_supports();
@@ -691,7 +691,7 @@ void sbbs_t::insert_indicator(void)
 	if(term_supports(ANSI)) {
 		ansi_save();
 		ansi_gotoxy(cols,1);
-		uchar z=curatr;                       /* and go to EOL */
+		uint z=curatr;                       /* and go to EOL */
 		if(console&CON_INSERT) {
 			attr(BLINK|BLACK|(LIGHTGRAY<<4));
 			outchar('I');
